# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "nadeemshaik1171998@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/addCollaboratorByNoteId": {
            "post": {
                "summary": "Add Colaborator to specific Note ",
                "description": " Add Colaborator to specific Note ",
                "operationId": "980d942fbef886155f252a99efe2f62a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "note_id"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "note_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collaborator created Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/updateNoteByCollaborator": {
            "post": {
                "summary": "Edit the note through Colaborator ",
                "description": " Edit the note through Colaborator",
                "operationId": "7360a7b63c332d5ea017e7caac45dfc0",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "note_id",
                                    "description"
                                ],
                                "properties": {
                                    "note_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note updated Sucessfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/removeCollaborator": {
            "post": {
                "summary": "Remove Colaborator from specific Note ",
                "description": " Remove Colaborator from specific Note ",
                "operationId": "0ece06f261e90482228201a9bd103a8a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "note_id"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "note_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collaborator deleted Sucessfully"
                    },
                    "404": {
                        "description": "Collaborater Not created"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/getAllCollaborators": {
            "get": {
                "summary": "Display all Colaborators",
                "description": " Display all Colaborators ",
                "operationId": "c91bdd51cfdc430e5ea6d642e10250ef",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "Fetched Collaborators Successfully"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}